Basic pid control working. Changed messages to float64 to work with rospid. Added depth_stamped message for Vector messages, depth for float64.

+++ float_control/nodes/static_command.py
@@ -4,11 +4,11 @@ import time
 import argparse
 import sys
 import rospy
-from std_msgs.msg import Float32
+from std_msgs.msg import Float32, Float64
 
 if __name__ == "__main__":
     rospy.init_node('static_command')
-    thruster_pub = rospy.Publisher('/thruster_cmd', Float32)
+    thruster_pub = rospy.Publisher('/thruster_cmd', Float64)
 
     r = rospy.Rate(5)  # 5 Hz
 
@@ -16,7 +16,7 @@ if __name__ == "__main__":
 
     while not rospy.is_shutdown():
         set_command = rospy.get_param("set_thruster_command", 0.0)
-        msg = Float32()
+        msg = Float64()
         msg.data = set_command
         thruster_pub.publish(msg)
