+++ float_control/nodes/static_depth.py
@@ -4,11 +4,11 @@ import time
 import argparse
 import sys
 import rospy
-from std_msgs.msg import Float32
+from std_msgs.msg import Float32, Float64
 
 if __name__ == "__main__":
     rospy.init_node('static_depth')
-    depth_pub = rospy.Publisher('/depth_target', Float32)
+    depth_pub = rospy.Publisher('/depth_target', Float64)
 
     r = rospy.Rate(5)  # 5 Hz
 
@@ -16,7 +16,7 @@ if __name__ == "__main__":
 
     while not rospy.is_shutdown():
         set_depth = rospy.get_param("set_target_depth", 0.0)
-        msg = Float32()
+        msg = Float64()
         msg.data = set_depth
         depth_pub.publish(msg)
